refactor: improve robustness and security of resource handling and parsing
This commit implements several critical improvements to prevent undefined behavior, integer overflows, and potential security vulnerabilities related to file I/O and memory management.

Key changes include:
- **Security & Robustness**: Added `MAX_SHADER_FILE_SIZE` limits to all shader-related file reads (cache, replacements, includes) to prevent multi-gigabyte allocation attacks or crashes due to `GetFileSize` failures.
- **Integer Overflow Prevention**: Migrated time-related calculations (e.g., `gTime`, `ticks_at_launch`, `OverlayNotice` timestamps) from 32-bit `GetTickCount` to 64-bit `GetTickCount64` to prevent wrapping after ~49.7 days. Applied 64-bit arithmetic to resource offset and size calculations to prevent overflows.
- **Memory Safety**:
    - Fixed `ResourceHandleInfo::SetDataCache` to perform a deep copy of data instead of adopting a pointer, ensuring validity after D3D runtime `Unmap`.
    - Improved validation of DXBC container structures in the shader assembler/disassembler.
    - Added bounds checking and path-boundary awareness for DLL loading and directory containment checks.
- **Input Validation**: Added validation for `ini` configuration values (e.g., `width_multiply`, `height_multiply`, `override_vertex_count`) to catch non-finite, negative, or excessively large values at parse time.
- **Compliance & Documentation**:
    - Added `SECURITY.md`, `THIRD_PARTY_NOTICES.md`, and `pcre2/LICENCE` to comply with redistribution requirements.
    - Updated GitHub release workflows to include necessary license and provenance notices in distribution zips.
    - Updated `CHANGELOG.md` and documentation.

+++ DirectX11/ShaderRegex.cpp
@@ -679,7 +679,9 @@ bool get_shader_model_from_bytecode(const void* data, size_t size, std::string* 
 	{
 		uint32_t offset = chunkOffsets[i];
 
-		if (offset + 12 > size)
+		// 64-bit arithmetic: a 32-bit offset + 12 can wrap to a small
+		// value and pass the check with a ~4GB out-of-bounds read:
+		if ((uint64_t)offset + 12 > size)
 			continue;
 
 		const uint8_t* chunk = buffer + offset;
